Further work on BlockStorage
- added source code file in src/PAL
- left initialization (because it's target specific) in Target folder
- moved common declarations and defines to global header
- removed function pointers from structure (don't make sense in C implementation)
- update code accordingly
